Move form implementation to a separate hook in auth
